% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFunction.R
\name{Cube}
\alias{Cube}
\title{Cubic Function}
\usage{
Cube(x)
}
\arguments{
\item{x}{numeric variable}
}
\value{
x^3
}
\description{
Takes numeric in puts x
}
\examples{
Cube(2)
}
\author{
Jishnu
}
