% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFunction.R
\name{Power}
\alias{Power}
\title{Power Function}
\usage{
Power(x, y)
}
\arguments{
\item{x}{numeric variable}

\item{y}{numeric variable}
}
\value{
x^y
}
\description{
Takes numeric in puts x & y
}
\examples{
Power(2,3)
}
\author{
Jishnu
}
